examples/secrets-init/source/requirements.txt
boto3
jproperties
pyyaml
toml
unflatten
ec2-metadata